add character position tracker
