+++ src/whylogs/core/statistics/__init__.py
@@ -5,10 +5,12 @@ from .counterstracker import CountersTracker
 from .numbertracker import NumberTracker
 from .schematracker import SchemaTracker
 from .thetasketch import ThetaSketch
+from .stringtracker import StringTracker
 
 __ALL__ = [
     CountersTracker,
     NumberTracker,
     SchemaTracker,
     ThetaSketch,
+    StringTracker
 ]
