+++ src/whylogs/core/statistics/datatypes/__init__.py
@@ -3,12 +3,10 @@ Define classes for tracking statistics for various data types
 """
 from .floattracker import FloatTracker
 from .integertracker import IntTracker
-from .stringtracker import StringTracker
 from .variancetracker import VarianceTracker
 
 __ALL__ = [
     FloatTracker,
     IntTracker,
-    StringTracker,
     VarianceTracker,
 ]
